moved proxy into model.py
